Function libnghttp2_sys::nghttp2_session_server_new3

pub unsafe extern "C" fn nghttp2_session_server_new3(
    session_ptr: *mut *mut nghttp2_session, 
    callbacks: *const nghttp2_session_callbacks, 
    user_data: *mut c_void, 
    option: *const nghttp2_option, 
    mem: *mut nghttp2_mem
) -> c_int

@function

Like nghttp2_session_server_new2(), but with additional custom memory allocator specified in the |mem|.

The |mem| can be NULL and the call is equivalent to nghttp2_session_server_new2().

This function does not take ownership |mem|. The application is responsible for freeing |mem|.

The library code does not refer to |mem| pointer after this function returns, so the application can safely free it.

This function returns 0 if it succeeds, or one of the following negative error codes:

:enum:NGHTTP2_ERR_NOMEM Out of memory.